feat: finish CabangModel and Create Cabang Producer API

+++ siretail/src/main/java/apap/tugasakhir/siretail/service/UserServiceImpl.java
@@ -64,7 +64,7 @@ public class UserServiceImpl implements UserService {
     }
 
     @Override
-    public UserModel getUserById(Long id) {
+    public UserModel getUserById(Integer id) {
         Optional<UserModel> user = userDb.findById(id);
         if (user.isPresent()){
             return user.get();
